Add memory barrier to our implementation of memzero (taken from BoringSSL).
The existing code uses the volatile keyword to try to get the compiler to not optimize away the memory wiping loop.  This should work on most compilers, though I'm not convinced it will work on all.  The other alternative, used in the Linux kernel, Google's BoringSSL, and Bitcoin Core is to use assembly intrinsics to issue a memory barrier.  This forces the C++ compiler (which lacks architectural knowledge) to back off.  Both approaches are complementary; this additional code gives us strictly greater assurance that the memory wipe is actually performed.

Note that this particular fix was originally copied from BoringSSL, which was released by Google under the ISC license, which is compatible with bc-crypto-base's BSD license.